Fix Analyzer Errors
This commit fixes various analyzer errors which include:
- fixing type casts (converting from num to int)
- fixing mismatched types in mock (forcing the correct type)

+++ lib/src/browser/hash_transition_manager.dart
@@ -69,7 +69,7 @@ class HashTransitionManager<T extends HashHistory>
     if (delta == 0) {
       return;
     }
-    int nextDomCheck = (_domCheck + delta).clamp(0, 2);
+    int nextDomCheck = (_domCheck + delta).clamp(0, 2).toInt();
     if (_domCheck == 0 && nextDomCheck != 0) {
       window.addEventListener('hashchange', _hashChangeHandler);
     } else if (_domCheck != 0 && nextDomCheck == 0) {
